src/main/java/org/example/Radio.java
package org.example;

public class Radio {
    private int currentStation;
    private int currentVolume;

    public int getCurrentStation() {
        return currentStation;
    }

    public int getCurrentVolume() {
        return currentVolume;
    }

    public void setCurrentStation(int newCurrentStation) {
        if (newCurrentStation > 9)
            currentStation = 0;
        else if (newCurrentStation < 0)
            currentStation = 9;
        else {
            currentStation = newCurrentStation;
        }
    }

    public void setCurrentVolume(int newCurrentVolume) {
        if (newCurrentVolume < 1) {
            return;
        }
        if (newCurrentVolume > 10) {
            return;
        }
        currentVolume = newCurrentVolume;
    }

    public void setNextStation() {
        int target = currentStation + 1;
        setCurrentStation(target);
    }

    public void setPrevStation() {
        int target = currentStation - 1;
        setCurrentStation(target);
    }

    public void increaseVolume() {
        if (currentVolume < 10) {
            currentVolume = currentVolume + 1;
        }
    }

    public void decreaseVolume() {
        if (currentVolume > 0) {
            currentVolume = currentVolume - 1;
        }
    }
}